@@ -25,16 +25,14 @@ namespace Vault.Models
 
         public string UserID { get; set; }
 
-        [DisplayName("* Description")]
+        [DisplayName("Description")]
         public string Description { get; set; }
 
         public string Username { get; set; }
 
         public string Password { get; set; }
 
-#pragma warning disable CA1056 // Uri properties should not be strings
         public string Url { get; set; }
-#pragma warning restore CA1056 // Uri properties should not be strings
 
         [DisplayName("Custom 1 Label")]
         public string UserDefined1Label { get; set; }
