@@ -5,9 +5,9 @@ namespace Vault.Models
     public class LoginViewModel
     {
         [Display(Name = "Username")]
-        public string UN1209 { get; set; }
+        public string Username { get; set; }
 
         [Display(Name = "Password")]
-        public string PW9804 { get; set; }
+        public string Password { get; set; }
     }
 }
