@@ -3,16 +3,16 @@
     public static class SqlStatements
     {
         public const string SelectSummary =
-            @"SELECT 
-                  CredentialID, 
-                  UserID, 
-                  Description, 
-                  Username, 
-                  Password, 
-                  Url 
-              FROM 
-                  Credentials 
-              WHERE 
+            @"SELECT
+                  CredentialID,
+                  UserID,
+                  Description,
+                  Username,
+                  Password,
+                  Url
+              FROM
+                  Credentials
+              WHERE
                   UserID = @UserID";
 
         public const string Select =
@@ -22,51 +22,51 @@
             "SELECT * FROM Credentials WHERE CredentialID = @CredentialID";
 
         public const string Insert =
-            @"INSERT INTO 
+            @"INSERT INTO
                   Credentials (
-                      CredentialID, 
-                      UserID, 
-                      Description, 
-                      Username, 
-                      Password, 
-                      Url, 
-                      UserDefined1Label, 
-                      UserDefined1, 
-                      UserDefined2Label, 
-                      UserDefined2, 
-                      Notes, 
+                      CredentialID,
+                      UserID,
+                      Description,
+                      Username,
+                      Password,
+                      Url,
+                      UserDefined1Label,
+                      UserDefined1,
+                      UserDefined2Label,
+                      UserDefined2,
+                      Notes,
                       PwdOptions
-                  ) 
+                  )
               VALUES (
-                  @CredentialID, 
-                  @UserID, 
-                  @Description, 
-                  @Username, 
-                  @Password, 
-                  @Url, 
-                  @UserDefined1Label, 
-                  @UserDefined1, 
-                  @UserDefined2Label, 
-                  @UserDefined2, 
-                  @Notes, 
+                  @CredentialID,
+                  @UserID,
+                  @Description,
+                  @Username,
+                  @Password,
+                  @Url,
+                  @UserDefined1Label,
+                  @UserDefined1,
+                  @UserDefined2Label,
+                  @UserDefined2,
+                  @Notes,
                   @PwdOptions
               )";
 
         public const string Update =
-            @"UPDATE 
-                  Credentials 
-              SET 
-                  Description = @Description, 
-                  Username = @Username, 
-                  Password = @Password, 
-                  Url = @Url, 
-                  UserDefined1Label = @UserDefined1Label, 
-                  UserDefined1 = @UserDefined1, 
-                  UserDefined2Label = @UserDefined2Label, 
-                  UserDefined2 = @UserDefined2, 
-                  Notes = @Notes, 
-                  PwdOptions = @PwdOptions 
-              WHERE 
+            @"UPDATE
+                  Credentials
+              SET
+                  Description = @Description,
+                  Username = @Username,
+                  Password = @Password,
+                  Url = @Url,
+                  UserDefined1Label = @UserDefined1Label,
+                  UserDefined1 = @UserDefined1,
+                  UserDefined2Label = @UserDefined2Label,
+                  UserDefined2 = @UserDefined2,
+                  Notes = @Notes,
+                  PwdOptions = @PwdOptions
+              WHERE
                   CredentialID = @CredentialID";
 
         public const string Delete =
