@@ -7,42 +7,42 @@
 
 @section main {
 
-    <div id="generate-form-dialog" class="modal in" style="display: block;">
-        <div class="modal-dialog modal-lg">
+    <div id="generate-form-dialog" class="modal show" style="display: block;">
+        <div class="modal-dialog modal-lg modal-dialog-centered">
             <div class="modal-content">
-                <form id="generate-form" action="" method="post" class="form-horizontal">
+                <form id="generate-form" action="" method="post">
                     <div class="modal-header">
                         <h4 class="modal-title">Generate Credentials</h4>
                     </div>
                     <div class="modal-body">
-                        <div class="form-group">
-                            <label for="UserID" class="col-sm-8 control-label">User ID</label>
-                            <div class="col-sm-16">
+                        <div class="row mb-3">
+                            <label for="UserID" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)">User ID</label>
+                            <div class="col-md-@(FieldColumnWidthMD)">
                                 <input type="text" size="40" id="UserID" name="UserID" value="@Model.UserID" class="form-control" />
                             </div>
                         </div>
                         <p>Enter a username and password to hash:</p>
-                        <div class="form-group">
-                            <label for="Username" class="col-sm-8 control-label">Username</label>
-                            <div class="col-sm-16">
+                        <div class="row mb-3">
+                            <label for="Username" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)">Username</label>
+                            <div class="col-md-@(FieldColumnWidthMD)">
                                 <input type="text" size="40" id="Username" name="Username" value="" class="form-control" />
                             </div>
                         </div>
-                        <div class="form-group">
-                            <label for="Password" class="col-sm-8 control-label">Password</label>
-                            <div class="col-sm-16">
+                        <div class="row mb-3">
+                            <label for="Password" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)">Password</label>
+                            <div class="col-md-@(FieldColumnWidthMD)">
                                 <input type="text" size="40" id="Password" name="Password" value="" class="form-control" />
                             </div>
                         </div>
-                        <div class="form-group">
-                            <label for="HashedUsername" class="col-sm-8 control-label">Hashed Username</label>
-                            <div class="col-sm-16">
+                        <div class="row mb-3">
+                            <label for="HashedUsername" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)">Hashed Username</label>
+                            <div class="col-md-@(FieldColumnWidthMD)">
                                 <input type="text" size="40" id="HashedUsername" name="HashedUsername" value="" class="form-control" />
                             </div>
                         </div>
-                        <div class="form-group">
-                            <label for="HashedPassword" class="col-sm-8 control-label">Hashed Password</label>
-                            <div class="col-sm-16">
+                        <div class="row mb-3">
+                            <label for="HashedPassword" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)">Hashed Password</label>
+                            <div class="col-md-@(FieldColumnWidthMD)">
                                 <input type="text" size="40" id="HashedPassword" name="HashedPassword" value="" class="form-control" />
                             </div>
                         </div>
@@ -59,8 +59,8 @@
 
 @section foot {
 
-    <script src="~/js/dist/generatevaultcredential.js" asp-append-version="true" integrity="sha512-wS+PSoLTK/fgM+49k9u+NeiII5vSDOkGk5RKqAGD5J0LiTZi9ZWlHV8GzNe1wBXkKELi6ok71zw8QFmytmxFzQ=="></script>
+    <script src="~/js/dist/GenerateVaultCredential.min.js" asp-append-version="true" integrity="sha512-muT2OhPLoCPfAh9UhyL00/0+4MsyGppBRdD5oVfIdOhSnyYSwA6PWH9IQUBjetqVnzcR1/3YyDOWS6/hj1OKZA=="></script>
 
-    <div class="modal-backdrop in"></div>
+    <div class="modal-backdrop show"></div>
 
 }
