@@ -20,6 +20,6 @@
         var _VAULT_GLOBALS = @Html.Raw(Model.AsJson());
     </script>
 
-    <script src="~/js/dist/index.js" asp-append-version="true" integrity="sha512-RF7j9K/Eqdb1qYRyTLrEtDnnk/p1gcqs4fckwRlxHpgzajNxgjpZd8aApP5jE1rhUrbYOsdsFnjkpnRMgJQ8Sw=="></script>
+    <script src="~/js/dist/Index.min.js" asp-append-version="true" integrity="sha512-8ni6tCdvJN9w0ilCAW3GJsa1yt/6M6oU7aPzRuXdS4GKtIl7/cI6PDELg+Lcqx7nDXCkTadP3X8/tbg2Zf1jrw=="></script>
 
 }
