@@ -1,88 +1,107 @@
 ﻿@model Vault.Models.Credential
 
-@using (Html.BeginForm("Update", "Main", FormMethod.Post, new { id = "credential-form", @class = "form-horizontal", autocomplete = "off" })) {
+<form asp-action="Update" asp-controller="Main" method="post" id="credential-form" class="needs-validation" novalidate>
 
-    <!-- These fake fields are a workaround for Chrome not respecting the autocomplete attribute and autofilling fields with the wrong values... -->
-    <input class="chrome-autocomplete-fake" type="text" name="chromeusername" />
-    <input class="chrome-autocomplete-fake" type="password" name="chromepassword" />
+    <div class="row mb-md-@(RowMarginMD)">
+        <label asp-for="Description" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-@(FieldColumnWidthMD)">
+            <input asp-for="Description" type="text" class="form-control" autocomplete="off" required />
+            <div class="invalid-feedback">Please enter a description</div>
+        </div>
+    </div>
 
-    <div class="form-group">
-        @Html.LabelFor(x => x.Description, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-16">
-            @Html.TextBoxFor(x => x.Description, new { @class = "form-control", autocomplete = "vault-desc" })
+    <div class="row mb-md-@(RowMarginMD)">
+        <label asp-for="Username" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-@(FieldColumnWidthMD)">
+            <input asp-for="Username" type="text" class="form-control" />
         </div>
     </div>
-    <div class="form-group">
-        @Html.LabelFor(x => x.Username, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-16">
-            @Html.TextBoxFor(x => x.Username, new { @class = "form-control", autocomplete = "vault-un" })
-        </div>
-    </div>
-    <div class="form-group">
-        @Html.LabelFor(x => x.Password, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-10">
-            @Html.TextBoxFor(x => x.Password, new { @class = "form-control", autocomplete = "vault-pw" })
-            <div class="password-strength"><span></span><div></div></div>
-        </div>
-        <div class="col-sm-6">
-            <button class="btn btn-default generate-password" tabindex="-1">Generate</button>
-            <a href="#" class="generate-password-options-toggle" tabindex="-1">Options</a>
-        </div>
-    </div>
-    <div class="form-group generate-password-options">
-        <label class="col-sm-8 control-label">Password Chars</label>
-        <div class="col-sm-16">
-            <div class="form-group row">
-                <div class="col-sm-6">
-                    <input class="form-control generate-password-option" id="len" name="len" value="16"/>
-                    @Html.HiddenFor(x => x.PwdOptions)
+
+    <div class="row mb-md-@(RowMarginMD)">
+        <label asp-for="Password" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-5">
+            <input asp-for="Password" type="text" class="form-control" autocomplete="off" />
+            <div class="password-strength">
+                <div class="progress">
+                    <div class="progress-bar" role="progressbar"></div>
                 </div>
-                <div class="col-sm-18 checkbox">
-                    <label><input type="checkbox" class="generate-password-option" id="ucase" name="ucase" value="1" checked="checked" /> A-Z</label> &nbsp;
-                    <label><input type="checkbox" class="generate-password-option" id="lcase" name="lcase" value="1" checked="checked" /> a-z</label> &nbsp;
-                    <label><input type="checkbox" class="generate-password-option" id="nums" name="nums" value="1" checked="checked" /> 0-9</label> &nbsp;
-                    <label><input type="checkbox" class="generate-password-option" id="symb" name="symb" value="1" checked="checked" /> Symbols</label>
-                </div>
+                <div class="status"></div>
+            </div>
+        </div>
+        <div class="col-md-4 mt-2 mt-md-0">
+            <button class="btn btn-outline-secondary generate-password" tabindex="-1">Generate</button>
+            <a href="#" class="btn btn-link generate-password-options-toggle" tabindex="-1">Options</a>
+        </div>
+    </div>
+
+    <div class="row mb-md-@(RowMarginMD) generate-password-options d-none">
+        <label class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)">Password Chars</label>
+        <div class="col-md-3">
+            <input type="number" class="form-control generate-password-option" id="len" name="len" value="16" />
+            <input asp-for="PwdOptions" type="hidden" />
+        </div>
+        <div class="col-md-6 d-flex align-items-center">
+            <div class="form-check me-3">
+                <input class="form-check-input generate-password-option" type="checkbox" id="ucase" name="ucase" value="1" checked>
+                <label class="form-check-label" for="ucase">A-Z</label>
+            </div>
+            <div class="form-check me-3">
+                <input class="form-check-input generate-password-option" type="checkbox" id="lcase" name="lcase" value="1" checked>
+                <label class="form-check-label" for="lcase">a-z</label>
+            </div>
+            <div class="form-check me-3">
+                <input class="form-check-input generate-password-option" type="checkbox" id="nums" name="nums" value="1" checked>
+                <label class="form-check-label" for="nums">0-9</label>
+            </div>
+            <div class="form-check">
+                <input class="form-check-input generate-password-option" type="checkbox" id="symb" name="symb" value="1" checked>
+                <label class="form-check-label" for="symb">Symbols</label>
             </div>
         </div>
     </div>
-    <div class="form-group">
-        @Html.LabelFor(x => x.Url, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-16">
-            @Html.TextBoxFor(x => x.Url, new { @class = "form-control", autocomplete = "vault-url" })
-        </div>
-    </div>
-    <div class="form-group">
-        @Html.LabelFor(x => x.UserDefined1Label, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-16">
-            @Html.TextBoxFor(x => x.UserDefined1Label, new { @class = "form-control", autocomplete = "vault-ul1" })
-        </div>
-    </div>
-    <div class="form-group">
-        @Html.LabelFor(x => x.UserDefined1, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-16">
-            @Html.TextBoxFor(x => x.UserDefined1, new { @class = "form-control", autocomplete = "vault-ud1" })
-        </div>
-    </div>
-    <div class="form-group">
-        @Html.LabelFor(x => x.UserDefined2Label, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-16">
-            @Html.TextBoxFor(x => x.UserDefined2Label, new { @class = "form-control", autocomplete = "vault-ul2" })
-        </div>
-    </div>
-    <div class="form-group">
-        @Html.LabelFor(x => x.UserDefined2, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-16">
-            @Html.TextBoxFor(x => x.UserDefined2, new { @class = "form-control", autocomplete = "vault-ud2" })
-        </div>
-    </div>
-    <div class="form-group">
-        @Html.LabelFor(x => x.Notes, new { @class = "col-sm-8 control-label" })
-        <div class="col-sm-16">
-            @Html.TextAreaFor(x => x.Notes, new { @class = "form-control", rows = "6" })
+
+    <div class="row mb-md-@(RowMarginMD)">
+        <label asp-for="Url" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-@(FieldColumnWidthMD)">
+            <input asp-for="Url" type="text" class="form-control" autocomplete="off" />
         </div>
     </div>
 
-    @Html.HiddenFor(x => x.CredentialID)
+    <div class="row mb-md-@(RowMarginMD)">
+        <label asp-for="UserDefined1Label" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-@(FieldColumnWidthMD)">
+            <input asp-for="UserDefined1Label" type="text" class="form-control" autocomplete="off" />
+        </div>
+    </div>
 
-}
+    <div class="row mb-md-@(RowMarginMD)">
+        <label asp-for="UserDefined1" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-@(FieldColumnWidthMD)">
+            <input asp-for="UserDefined1" type="text" class="form-control" autocomplete="off" />
+        </div>
+    </div>
+
+    <div class="row mb-md-@(RowMarginMD)">
+        <label asp-for="UserDefined2Label" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-@(FieldColumnWidthMD)">
+            <input asp-for="UserDefined2Label" type="text" class="form-control" autocomplete="off" />
+        </div>
+    </div>
+
+    <div class="row mb-md-@(RowMarginMD)">
+        <label asp-for="UserDefined2" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-@(FieldColumnWidthMD)">
+            <input asp-for="UserDefined2" type="text" class="form-control" autocomplete="off" />
+        </div>
+    </div>
+
+    <div class="row">
+        <label asp-for="Notes" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+        <div class="col-md-@(FieldColumnWidthMD)">
+            <textarea asp-for="Notes" class="form-control" rows="6"></textarea>
+        </div>
+    </div>
+
+    <input asp-for="CredentialID" type="hidden" />
+
+</form>
