@@ -5,6 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, height=device-height, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
     <title>@ViewBag.Title</title>
+    <link rel="prefetch" href="~/css/dist/fonts/bootstrap-icons.woff2" as="font" type="font/woff2" crossorigin>
     <meta name="robots" content="noindex">
     <meta name="description" content="Vault Password Manager">
     <link rel="apple-touch-icon" sizes="57x57" href="~/img/apple-touch-icon-57x57.png">
@@ -29,9 +30,8 @@
     <meta name="msapplication-TileImage" content="~/img/mstile-144x144.png">
     <meta name="msapplication-config" content="~/img/browserconfig.xml">
     <meta name="theme-color" content="#7663de">
-    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css">
-    <link rel="stylesheet" href="~/css/external/bootstrap.css" />
-    <link rel="stylesheet" href="~/css/main.css" />
+    <link rel="stylesheet" href="~/css/dist/bootstrap-icons.min.css" />
+    <link rel="stylesheet" href="~/css/dist/main.min.css" />
     @RenderSection("head", false)
 </head>
 <body class="@(ViewBag.BodyClass ?? "base")">
