@@ -1,37 +1,33 @@
 ﻿@model Vault.Models.LoginViewModel
 
 <div id="login-form-dialog" class="modal">
-    <div class="modal-valign-container">
-        <div class="modal-dialog modal-lg modal-valign">
-            <div class="modal-content">
-                @using (Html.BeginForm("Login", "Main", FormMethod.Post, new { id = "login-form", @class = "form-horizontal" })) {
+    <div class="modal-dialog modal-lg modal-dialog-centered">
+        <div class="modal-content">
+            <form id="login-form" asp-action="Login" asp-controller="Home" method="post" class="needs-validation" novalidate>
                 <div class="modal-header">
                     <h4 class="modal-title">Log In</h4>
                 </div>
                 <div class="modal-body">
                     <p>Enter your username and master password:</p>
-                    <!-- These fake fields are a workaround for Chrome not respecting the autocomplete attribute and autofilling fields with the wrong values... -->
-                    <input class="chrome-autocomplete-fake" type="text" name="chromeusername"/>
-                    <input class="chrome-autocomplete-fake" type="password" name="chromepassword"/>
-                    <div class="form-group">
-                        @Html.LabelFor(model => model.UN1209, new { @class = "col-sm-8 control-label" })
-                        <div class="col-sm-16">
-                            @Html.PasswordFor(model => model.UN1209, new { @class = "form-control", autocomplete = "un1209" })
+                    <div class="row mb-md-@(RowMarginMD)">
+                        <label asp-for="Username" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+                        <div class="col-md-@(FieldColumnWidthMD)">
+                            <input asp-for="Username" type="text" class="form-control" required />
+                            <div class="invalid-feedback">Please enter a username</div>
                         </div>
                     </div>
-                    <div class="form-group">
-                        @Html.LabelFor(model => model.PW9804, new { @class = "col-sm-8 control-label" })
-                        <div class="col-sm-16">
-                            @Html.PasswordFor(model => model.PW9804, new { @class = "form-control", autocomplete = "pw9804" })
+                    <div class="row">
+                        <label asp-for="Password" class="col-md-@(LabelColumnWidthMD) col-form-label @(LabelVisualEffectClasses)"></label>
+                        <div class="col-md-@(FieldColumnWidthMD)">
+                            <input asp-for="Password" type="password" class="form-control" required />
+                            <div class="invalid-feedback">Please enter a password</div>
                         </div>
                     </div>
                 </div>
                 <div class="modal-footer">
-                    <span class="validation-message"></span>
-                    <button type="submit" class="btn btn-primary">Log In</button>
+                    <span class="badge rounded-pill bg-danger me-3 d-none">Login Failed</span> <button type="submit" class="btn btn-primary">Log In</button>
                 </div>
-                }
-            </div>
+            </form>
         </div>
     </div>
 </div>
