@@ -1,37 +1,39 @@
-﻿<div id="spinner">
-    <img src="~/img/ajax-loader.gif" />
+﻿<div id="spinner" class="d-none">
+    <div class="spinner-border text-light" role="status">
+        <span class="visually-hidden">Loading...</span>
+    </div>
 </div>
 
-<div class="row" id="controls">
-    <div class="col-sm-16 relative-container">
-        <label class="sr-only" for="search">Search</label>
-        <input type="text" id="search" name="search" class="form-control" placeholder="Search..." autocomplete="off" />
-        <button id="clear-search" class="btn btn-link">
-            <span class="fa fa-times"></span>
+<div class="row d-none mt-3" id="controls">
+    <div class="col-md-8">
+        <label class="visually-hidden" for="search">Search</label>
+        <div class="search-input">
+            <input type="text" id="search" name="search" class="form-control" placeholder="Search..." autocomplete="off" />
+            <button id="clear-search" class="btn btn-link">
+                <span class="bi bi-x-lg"></span>
+            </button>
+        </div>
+    </div>
+    <div class="col-md-2 mt-2 mt-md-0">
+        <label class="visually-hidden" for="new">New Item</label>
+        <button id="new" class="btn btn-primary w-100">
+            <span class="bi bi-plus-lg"></span> New
         </button>
     </div>
-    <div class="col-sm-4">
-        <label class="sr-only" for="new">New Item</label>
-        <button id="new" class="btn btn-primary btn-block">
-            <span class="fa fa-plus"></span> New
-        </button>
-    </div>
-    <div class="col-sm-4 hidden-xs">
-        <label class="sr-only" for="admin">Admin</label>
-        <button id="admin" class="btn btn-default btn-block">
-            <i class="fa fa-cog"></i> Admin
+    <div class="col-md-2 mt-2 mt-md-0">
+        <label class="visually-hidden" for="admin">Admin</label>
+        <button id="admin" class="btn btn-outline-secondary w-100">
+            <i class="bi bi-gear"></i> Admin
         </button>
     </div>
 </div>
 
 <div class="row">
-    <div class="col-md-24" id="container"></div>
+    <div class="col-md-12 mt-3" id="container"></div>
 </div>
 
 <div id="modal" class="modal" tabindex="-1">
-    <div class="modal-valign-container">
-        <div class="modal-dialog modal-lg modal-valign">
-            <div class="modal-content" id="modal-content"></div>
-        </div>
+    <div class="modal-dialog modal-dialog-centered modal-lg">
+        <div class="modal-content" id="modal-content"></div>
     </div>
 </div>
